+++ packages/fabric/editor-core/__tests__/editor/plugins/media.ts
@@ -1,18 +1,27 @@
 import { name } from '../../../package.json';
 import { mediaPlugin } from '../../../src/editor/plugins';
+import { EditorPlugin } from '../../../src/editor/types';
+
+const getNodeNames = (plugin: EditorPlugin) =>
+  plugin.nodes ? plugin.nodes().map(node => node.name) : [];
 
 describe(name, () => {
   describe('Plugins -> Media', () => {
-    it('should be defined', () => {
-      expect(mediaPlugin).toBeDefined();
+    it('should not have mediaSingle node by default', () => {
+      const availableNodes = getNodeNames(mediaPlugin());
+      expect(availableNodes).toHaveLength(2);
+      expect(availableNodes).not.toContain('mediaSingle');
     });
 
-    it('should have primaryToolbarComponent', () => {
-      expect(mediaPlugin.primaryToolbarComponent).toBeDefined();
-    });
-
-    it('should have secondaryToolbarComponent', () => {
-      expect(mediaPlugin.secondaryToolbarComponent).toBeDefined();
+    it('should have mediaSingle node when allowMediaSingle is true', () => {
+      const availableNodes = getNodeNames(
+        mediaPlugin({
+          provider: Promise.resolve() as any,
+          allowMediaSingle: true,
+        }),
+      );
+      expect(availableNodes).toHaveLength(3);
+      expect(availableNodes).toContain('mediaSingle');
     });
   });
 });