+++ packages/fabric/editor-common/CHANGELOG.md
@@ -1,5 +1,9 @@
 # @atlaskit/editor-common
 
+## 5.0.6
+- [patch] Add default center layout support for single image [6113e02](https://bitbucket.org/atlassian/atlaskit-mk-2/commits/6113e02)
+- [patch] Add default center layout support for single image [6113e02](https://bitbucket.org/atlassian/atlaskit-mk-2/commits/6113e02)
+
 ## 5.0.5
 
 ## 5.0.4